+++ PennyAuctionBackend/Dtos/AuctionItem/PlaceBidRequest.cs
@@ -1,7 +1,13 @@
 namespace PennyAuctionBackend.Dtos.AuctionItem;
 
-public class PlaceBidRequest
-{
-    public int AuctionItemId { get; set; }
-    public int BidAmount { get; set; }
+public class PlaceBidRequest {
+	public int AuctionItemId {
+		get;
+		set;
+	}
+
+	public long BidAmount {
+		get;
+		set;
+	}
 }
